set "shell = True" when call subprocess.Popen
On unix, when "shell = False" (default), the args must be a list of
args. While on Windows, filling args with a list causes problems. As a
trade-off, set the this flag and using a string as the command, although
this might sacrifice some performance, but that's not the main concern
in this script.

Signed-off-by: Heikki Hokkanen <hoxu@users.sf.net>