Search '/usr/share/gitstats/' for static files as well.
The first path to be searched is the directory of the binary, then
/usr/share/gitstats/. This is so that you can have both a "global" gitstats
version and a user-specific checkout installed, and the static files should be
taken from the matching place.

This change is done in preparation for Debian packaging.

Vincent Fourmond <fourmond@debian.org> is planning to package gitstats for
debian (RFP #440392).