use python2 enviroment
On systems like Arch Linux `python` defaults to `python3` preventing
gitstats to start. This fixes it.

Signed-off-by: Heikki Hokkanen <hoxu@users.sf.net>

[hoxu@users.sf.net: Debian Jessie and CentOS 6.5 have /usr/bin/python2,
but OS X Yosemite only has /usr/bin/python2.7. There does not seem to be
a portable way to refer to python 2.x, so unfortunately on some
platforms the shebang needs to be modified manually]